fix: Move intern did not make original pair of removed intern into paired intern if left with only one person., fixed any underlying copatiility issues.
